Move bootstrap endpoint from api to frontend
